Rename pkg/metrics/config to pkg/metrics/metricsconfig
It's more convenient to import this package as metricsconfig than config, so
let's call it metricsconfig.

Technically it is a breaking change, but only developer facing.
